+++ midi-bt-receiver/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "FSharp.suggestGitignore": false
+}
