+++ midi-bt-receiver/platformio.ini
@@ -10,18 +10,16 @@
 
 [env:micro_guitar_receiver]
 platform = espressif32
-board = esp32dev
+board = arduino_nano_esp32
+;board = esp32dev
 framework = arduino
-upload_speed = 115200
-upload_port = /dev/cu.usbserial-0001
-board_upload.resetmethod = nodemcu
-monitor_speed = 115200
+upload_port = /dev/cu.usbmodemECDA3B64AB842
 lib_deps =
     arduino-libraries/ArduinoBLE
-;upload_port = /dev/cu.usbserial-14210
 extra_scripts = pre:device_setup_guitar.py
 build_flags = 
     -DDEVICE_NAME="\"MIDI guitar receiver\""
+upload_protocol = dfu
 
 [env:micro_bass_receiver]
 platform = espressif32
