+++ midi-bt-transmitter/platformio.ini
@@ -14,10 +14,7 @@ board = esp32dev
 framework = arduino
 lib_deps =
     arduino-libraries/ArduinoBLE
-upload_port = /dev/cu.usbserial-410
-upload_speed = 115200
-;board_upload.resetmethod = nodemcu
-monitor_speed = 115200
+upload_port = /dev/cu.usbserial-0001
 build_flags = 
     -DRECEIVER_DEVICE_NAME="\"MIDI guitar receiver\""
